chore: Move AssetType logic to its own component
