+++ packages/utils-general/src/blockchain-facade/Location.ts
@@ -1,0 +1,73 @@
+export const THAILAND_REGIONS_PROVINCES_MAP = {
+    Northeast: [
+        'Amnat Charoen',
+        'Bueng Kan, Buri Ram',
+        'Chaiyaphum, Kalasin',
+        'Khon Kaen, Loei',
+        'Maha Sarakham',
+        'Mukdahan',
+        'Nakhon Phanom',
+        'Nakhon Ratchasima',
+        'Nong Bua Lamphu',
+        'Nong Khai',
+        'Roi Et',
+        'Sakon Nakhon',
+        'Si Sa Ket',
+        'Surin',
+        'Ubon Ratchathani',
+        'Udon Thani',
+        'Yasothon'
+    ],
+    North: [
+        'Chiang Mai',
+        'Chiang Rai',
+        'Lampang',
+        'Lamphun',
+        'Mae Hong Son',
+        'Nan',
+        'Phayao',
+        'Phrae',
+        'Uttaradit'
+    ],
+    West: ['Tak', 'Kanchanaburi', 'Ratchaburi', 'Phetchaburi', 'Prachuap Khiri Khan'],
+    Central: [
+        'Sukhothai',
+        'Phitsanulok',
+        'Phichit',
+        'Kamphaeng Phet',
+        'Phetchabun',
+        'Nakhon Sawan',
+        'Uthai Thani',
+        'Ang Thong',
+        'Phra Nakhon Si Ayutthaya',
+        'Bangkok, Chai Nat',
+        'Lop Buri',
+        'Nakhon Pathom',
+        'Nonthaburi',
+        'Pathum Thani',
+        'Samut Prakan',
+        'Samut Sakhon',
+        'Samut Songkhram',
+        'Saraburi',
+        'Sing Buri',
+        'Suphan Buri',
+        'Nakhon Nayok'
+    ],
+    East: ['Chachoengsao', 'Chanthaburi', 'Chon Buri', 'Prachin Buri', 'Rayong', 'Sa Kaeo', 'Trat'],
+    South: [
+        'Chumphon',
+        'Nakhon Si Thammarat',
+        'Narathiwat',
+        'Pattani',
+        'Phatthalung',
+        'Songkhla',
+        'Surat Thani',
+        'Yala',
+        'Krabi',
+        'Phang Nga',
+        'Phuket',
+        'Ranong',
+        'Satun',
+        'Trang'
+    ]
+};
