+++ packages/utils-general/src/index.ts
@@ -1,21 +1,5 @@
-// Copyright 2018 Energy Web Foundation
-// This file is part of the Origin Application brought to you by the Energy Web Foundation,
-// a global non-profit organization focused on accelerating blockchain technology across the energy sector,
-// incorporated in Zug, Switzerland.
-//
-// The Origin Application is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-// This is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY and without an implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details, at <http://www.gnu.org/licenses/>.
-//
-// @authors: slock.it GmbH; Heiko Burkhardt, heiko.burkhardt@slock.it; Martin Kuechler, martin.kuchler@slock.it
 import * as BlockchainDataModelEntity from './blockchain-facade/BlockchainDataModelEntity';
 import * as Configuration from './blockchain-facade/Configuration';
-
 
 export { ContractEventHandler } from './blockchain-facade/ContractEventHandler';
 export { EventHandlerManager } from './blockchain-facade/EventHandlerManager';
@@ -24,6 +8,6 @@ export { Configuration };
 export { validateJson } from './off-chain-data/json-validator';
 export { TimeFrame, Currency, Compliance } from './blockchain-facade/EnumExports';
 export { deploy } from './deployment/deploy';
-export { IAssetService, IRECAssetService } from './blockchain-facade/AssetTypeService';
+export { IAssetService, IRECAssetService, DecodedAssetType, EncodedAssetType } from './blockchain-facade/AssetTypeService';
 export { extendArray } from './extensions/array.extensions';
-
+export { THAILAND_REGIONS_PROVINCES_MAP } from './blockchain-facade/Location';
